Move and publish some functions related to warnings.
